refactor: ProfileFollowButton을 공통 컴포넌트로 이동해서 사용
